Fixed linux compilation and removed couple warnings

+++ LibCarla/source/carla/geom/Mesh.cpp
@@ -308,7 +308,7 @@ namespace geom {
 
     const size_t vertex_to_start_concating = v_num - num_vertices_to_link;
    
-    for( int i = 1; i < num_vertices_to_link; ++i )
+    for( size_t i = 1; i < num_vertices_to_link; ++i )
     {
       _indexes.push_back( vertex_to_start_concating + i );
       _indexes.push_back( vertex_to_start_concating + i  + 1 );
